+++ __tests__app__/app/home.js
@@ -1,6 +1,6 @@
 import React from 'react';
 import {
-	Page, Button, Text, Loader, DataComponent, Utils, SocketRequest,
+	Page, Text, Loader, DataComponent, Utils, SocketRequest,
 } from '../../src/app';
 
 import './home.css';
@@ -145,17 +145,25 @@ export default class Home extends Page {
 					loaderSize="small"
 					tookDisabled
 				/>
-				<Button href="/about" id="navigate-button">About page</Button>
-				<Button href="/" id="refresh-button" refreshContent>Refresh content</Button>
-				<Button href="/test" id="test-button">Invalid page</Button>
-				<Button id="state-button" onClick={() => this.getContext().pushState(null, { test: 1 })}>Push state query</Button>
-				<Button
+				<button onClick={() => this.getContext().navigate('/about')} id="navigate-button" type="button">About page</button>
+				<button onClick={() => this.getContext().refreshContent()} id="refresh-button" type="button">Refresh content</button>
+				<button onClick={() => this.getContext().navigate('/test')} id="test-button" type="button">Invalid page</button>
+				<button onClick={() => this.getContext().navigate('/layout-test')} id="layout-button" type="button">Different layout</button>
+				<button
+					id="state-button"
+					onClick={() => this.getContext().pushState(null, { test: 1 })}
+					type="button"
+				>
+					Push state query
+				</button>
+				<button
 					onClick={async () => {
 						console.log(await this.call('user.getAsyncError'));
 					}}
+					type="button"
 				>
 					Async request
-				</Button>
+				</button>
 				<a href="/about" onClick={Utils.anchorNavigation}>About page anchor</a>
 				<div>
 					<Text tag="p" dictionaryKey="args" args={['one', 'two', 'three']} />