+++ src/app/router.js
@@ -30,24 +30,30 @@ class Router {
 	 */
 	getRoute() {
 		const p = this.parseUrl();
+		return this.findRoute(p.path);
+	}
+
+	findRoute(path) {
 		let route;
+		if (path.indexOf('http') === 0) {
+			const p = url.parse(path, true);
+			path = p.path;
+		}
 		Object.keys(this._routes).forEach((spec) => {
 			if (route) {
 				return;
 			}
 			const r = new RouteParser(spec);
-			const match = r.match(p.path);
-			if (match === false) {
+			if (r.match(path) === false) {
 				return;
 			}
 			route = this._routes[spec];
 		});
-
 		return route || null;
 	}
 
 	pushState(path = null, q = {}) {
-		const { query, pathname } = this.parseUrl();
+		const { pathname } = this.parseUrl();
 		if (!path) {
 			// eslint-disable-next-line no-param-reassign
 			path = pathname;
@@ -56,16 +62,24 @@ class Router {
 			history.pushState(null, null, path);
 			return;
 		}
-		Object.keys(q).forEach((key) => {
-			const value = q[key];
-			if (value === undefined) {
-				delete query[key];
-			} else {
-				query[key] = value;
-			}
-		});
+		const s = this.stringifyQuery(q);
+		history.pushState(null, null, s ? `${path}${s}` : path);
+	}
+
+	stringifyQuery(q = {}) {
+		const { query } = this.parseUrl();
+		if (q) {
+			Object.keys(q).forEach((key) => {
+				const value = q[key];
+				if (value === undefined) {
+					delete query[key];
+				} else {
+					query[key] = value;
+				}
+			});
+		}
 		const s = qs.stringify(query);
-		history.pushState(null, null, s ? `${path}?${s}` : path);
+		return s ? `?${s}` : null;
 	}
 
 	parseUrl(params = false) {
@@ -108,9 +122,10 @@ class Route {
 	 * @param {JSX.Element} route.component
 	 * @param {string} route.title
 	 * @param {Object.<string, any>} route.initialData
+	 * @param {string} route.layout
 	 */
 	static create(route) {
-		return new this(route.spec, route.component, route.title, route.initialData);
+		return new this(route.spec, route.component, route.title, route.initialData, route.layout);
 	}
 
 	/**
@@ -119,12 +134,14 @@ class Route {
 	 * @param {JSX.Element} component
 	 * @param {string} title
 	 * @param {Object.<string, any>} initialData
+	 * @param {string} layout
 	 */
-	constructor(spec, component, title, initialData = {}) {
+	constructor(spec, component, title, initialData = {}, layout = null) {
 		this.spec = spec;
 		this.component = component;
 		this.title = title;
 		this.initialData = initialData;
+		this.layout = layout;
 	}
 
 	getComponent() {